requirements.txt
Django == 5.06
django-crispy-forms == 2.1
crispy-bootstrap4 == 2024.1
Pillow == 10.3.0
djangorestframework>=3.12.4,<3.13
psycopg2-binary>=2.9, <2.10